requirements.txt
snowflake-connector-python
snowflake-ml-python==1.2.2
snowflake-snowpark-python
